[SOIN] Extrais des composants pour les différents blocs de préférences
